Refactor source files for the LoadBalancerModelGenerator
