+++ internal/controller/model_nodeport.go
@@ -1,13 +1,10 @@
 package controller
 
 import (
-	"fmt"
-
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	corelisters "k8s.io/client-go/listers/core/v1"
 
-	"github.com/cloudandheat/ch-k8s-lbaas/internal/config"
 	"github.com/cloudandheat/ch-k8s-lbaas/internal/model"
 	"github.com/cloudandheat/ch-k8s-lbaas/internal/openstack"
 )
@@ -18,23 +15,6 @@ type NodePortLoadBalancerModelGenerator struct {
 	nodes         corelisters.NodeLister
 }
 
-type ClusterIPLoadBalancerModelGenerator struct {
-	l3portmanager openstack.L3PortManager
-	services      corelisters.ServiceLister
-	nodes         corelisters.NodeLister
-}
-
-type PodLoadBalancerModelGenerator struct {
-	l3portmanager openstack.L3PortManager
-	services      corelisters.ServiceLister
-	nodes         corelisters.NodeLister
-	endpoints     corelisters.EndpointsLister
-}
-
-type LoadBalancerModelGenerator interface {
-	GenerateModel(portAssignment map[string]string) (*model.LoadBalancer, error)
-}
-
 func NewNodePortLoadBalancerModelGenerator(
 	l3portmanager openstack.L3PortManager,
 	services corelisters.ServiceLister,
@@ -43,54 +23,6 @@ func NewNodePortLoadBalancerModelGenerator(
 		l3portmanager: l3portmanager,
 		services:      services,
 		nodes:         nodes,
-	}
-}
-
-func NewClusterIPLoadBalancerModelGenerator(
-	l3portmanager openstack.L3PortManager,
-	services corelisters.ServiceLister,
-	nodes corelisters.NodeLister) *ClusterIPLoadBalancerModelGenerator {
-	return &ClusterIPLoadBalancerModelGenerator{
-		l3portmanager: l3portmanager,
-		services:      services,
-		nodes:         nodes,
-	}
-}
-
-func NewPodLoadBalancerModelGenerator(
-	l3portmanager openstack.L3PortManager,
-	services corelisters.ServiceLister,
-	nodes corelisters.NodeLister,
-	endpoints corelisters.EndpointsLister) *PodLoadBalancerModelGenerator {
-	return &PodLoadBalancerModelGenerator{
-		l3portmanager: l3portmanager,
-		services:      services,
-		nodes:         nodes,
-		endpoints:     endpoints,
-	}
-}
-
-func NewLoadBalancerModelGenerator(
-	backendLayer config.BackendLayer,
-	l3portmanager openstack.L3PortManager,
-	services corelisters.ServiceLister,
-	nodes corelisters.NodeLister,
-	endpoints corelisters.EndpointsLister) (LoadBalancerModelGenerator, error) {
-	switch backendLayer {
-	case config.BackendLayerNodePort:
-		return NewNodePortLoadBalancerModelGenerator(
-			l3portmanager, services, nodes,
-		), nil
-	case config.BackendLayerClusterIP:
-		return NewClusterIPLoadBalancerModelGenerator(
-			l3portmanager, services, nodes,
-		), nil
-	case config.BackendLayerPod:
-		return NewPodLoadBalancerModelGenerator(
-			l3portmanager, services, nodes, endpoints,
-		), nil
-	default:
-		return nil, fmt.Errorf("invalid backend type: %q", backendLayer)
 	}
 }
 
@@ -162,12 +94,4 @@ func (g *NodePortLoadBalancerModelGenerator) GenerateModel(portAssignment map[st
 	}
 
 	return result, nil
-}
-
-func (g *ClusterIPLoadBalancerModelGenerator) GenerateModel(portAssignment map[string]string) (*model.LoadBalancer, error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
-func (g *PodLoadBalancerModelGenerator) GenerateModel(portAssignment map[string]string) (*model.LoadBalancer, error) {
-	return nil, fmt.Errorf("not implemented")
 }
