+++ internal/controller/model_nodeport_test.go
@@ -16,7 +16,7 @@ import (
 	ostesting "github.com/cloudandheat/ch-k8s-lbaas/internal/openstack/testing"
 )
 
-type generatorFixture struct {
+type nodePortGeneratorFixture struct {
 	t *testing.T
 
 	l3portmanager *ostesting.MockL3PortManager
@@ -27,8 +27,8 @@ type generatorFixture struct {
 	kubeobjects   []runtime.Object
 }
 
-func newGeneratorFixture(t *testing.T) *generatorFixture {
-	f := &generatorFixture{}
+func newNodePortGeneratorFixture(t *testing.T) *nodePortGeneratorFixture {
+	f := &nodePortGeneratorFixture{}
 	f.t = t
 	f.l3portmanager = ostesting.NewMockL3PortManager()
 	f.serviceLister = []*corev1.Service{}
@@ -54,12 +54,12 @@ func newGeneratorFixture(t *testing.T) *generatorFixture {
 	return f
 }
 
-func (f *generatorFixture) addNode(node *corev1.Node) {
+func (f *nodePortGeneratorFixture) addNode(node *corev1.Node) {
 	f.nodeLister = append(f.nodeLister, node)
 	f.kubeobjects = append(f.kubeobjects, node)
 }
 
-func (f *generatorFixture) newGenerator() (*NodePortLoadBalancerModelGenerator, kubeinformers.SharedInformerFactory) {
+func (f *nodePortGeneratorFixture) newGenerator() (*NodePortLoadBalancerModelGenerator, kubeinformers.SharedInformerFactory) {
 	f.kubeclient = k8sfake.NewSimpleClientset(f.kubeobjects...)
 	k8sI := kubeinformers.NewSharedInformerFactory(f.kubeclient, noResyncPeriodFunc())
 	services := k8sI.Core().V1().Services()
@@ -81,12 +81,12 @@ func (f *generatorFixture) newGenerator() (*NodePortLoadBalancerModelGenerator, 
 	return g, k8sI
 }
 
-func (f *generatorFixture) addService(svc *corev1.Service) {
+func (f *nodePortGeneratorFixture) addService(svc *corev1.Service) {
 	f.serviceLister = append(f.serviceLister, svc)
 	f.kubeobjects = append(f.kubeobjects, svc)
 }
 
-func (f *generatorFixture) runWith(body func(g *NodePortLoadBalancerModelGenerator)) {
+func (f *nodePortGeneratorFixture) runWith(body func(g *NodePortLoadBalancerModelGenerator)) {
 	g, k8sI := f.newGenerator()
 	stopCh := make(chan struct{})
 	defer close(stopCh)
@@ -123,15 +123,15 @@ func anyPort(t *testing.T, items []model.PortForward, inboundPort int32, protoco
 	}, "no PortForward found for %d/%s in %#v", inboundPort, protocol, items)
 }
 
-func (f *generatorFixture) matchDestinationAddresses(p model.PortForward) {
+func (f *nodePortGeneratorFixture) matchDestinationAddresses(p model.PortForward) {
 	assert.Equal(f.t, len(f.nodeLister), len(p.DestinationAddresses))
 	for _, node := range f.nodeLister {
 		assert.Contains(f.t, p.DestinationAddresses, node.Status.Addresses[1].Address)
 	}
 }
 
-func TestReturnsEmptyModelForEmptyAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortReturnsEmptyModelForEmptyAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	f.runWith(func(g *NodePortLoadBalancerModelGenerator) {
 		m, err := g.GenerateModel(map[string]string{})
@@ -141,8 +141,8 @@ func TestReturnsEmptyModelForEmptyAssignment(t *testing.T) {
 	})
 }
 
-func TestReturnsEmptyModelForNilAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortReturnsEmptyModelForNilAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	f.runWith(func(g *NodePortLoadBalancerModelGenerator) {
 		m, err := g.GenerateModel(nil)
@@ -152,8 +152,8 @@ func TestReturnsEmptyModelForNilAssignment(t *testing.T) {
 	})
 }
 
-func TestSinglePortSingleServiceAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortSinglePortSingleServiceAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	svc := newService("svc-1")
 	svc.Spec.Ports = []corev1.ServicePort{
@@ -185,8 +185,8 @@ func TestSinglePortSingleServiceAssignment(t *testing.T) {
 	})
 }
 
-func TestSinglePortMultiServiceAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortSinglePortMultiServiceAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	svc1 := newService("svc-1")
 	svc1.Spec.Ports = []corev1.ServicePort{
@@ -238,8 +238,8 @@ func TestSinglePortMultiServiceAssignment(t *testing.T) {
 	})
 }
 
-func TestMultiPortSingleServiceAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortMultiPortSingleServiceAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	svc1 := newService("svc-1")
 	svc1.Spec.Ports = []corev1.ServicePort{
@@ -296,8 +296,8 @@ func TestMultiPortSingleServiceAssignment(t *testing.T) {
 	})
 }
 
-func TestMultiPortMultiServiceAssignment(t *testing.T) {
-	f := newGeneratorFixture(t)
+func TestNodePortMultiPortMultiServiceAssignment(t *testing.T) {
+	f := newNodePortGeneratorFixture(t)
 
 	svc1 := newService("svc-1")
 	svc1.Spec.Ports = []corev1.ServicePort{
